created PostsPage.js and rendered the information on App,js
